Bug fixes and functionality
Changed Copy+Next to only be for the current fish
Also changed the way that fish are inserted into the list in order to
minimize confusion.  Fish are still sorted by frame they are added, but
the IDs correspond to the order in which the fish are added.  This may
need to be changed in the future.